2.0.5 fix for timeout for count

+++ HydroserverToolsBusinessObjects/DatabaseStatistics.cs
@@ -10,21 +10,21 @@ namespace HydroserverToolsBusinessObjects.Models
     public class DatabaseTableValueCountModel
     {
         [DisplayName("Sites")]
-        public int SiteCount { get; set; }
+        public Int64 SiteCount { get; set; }
         [DisplayName("Variables")]
-        public int VariablesCount { get; set; }
-        public int OffsetTypesCount { get; set; }
-        public int SourcesCount { get; set; }
-        public int MethodsCount { get; set; }
-        public int LabMethodsCount { get; set; }
-        public int SamplesCount { get; set; }
-        public int QualifiersCount { get; set; }
-        public int QualityControlLevelsCount { get; set; }
-        public int DataValuesCount { get; set; }
-        public int GroupDescriptionsCount { get; set; }
-        public int GroupsCount { get; set; }
-        public int DerivedFromCount { get; set; }
-        public int CategoriesCount { get; set; }
-        public int SeriesCatalog { get; set; }
+        public Int64 VariablesCount { get; set; }
+        public Int64 OffsetTypesCount { get; set; }
+        public Int64 SourcesCount { get; set; }
+        public Int64 MethodsCount { get; set; }
+        public Int64 LabMethodsCount { get; set; }
+        public Int64 SamplesCount { get; set; }
+        public Int64 QualifiersCount { get; set; }
+        public Int64 QualityControlLevelsCount { get; set; }
+        public Int64 DataValuesCount { get; set; }
+        public Int64 GroupDescriptionsCount { get; set; }
+        public Int64 GroupsCount { get; set; }
+        public Int64 DerivedFromCount { get; set; }
+        public Int64 CategoriesCount { get; set; }
+        public Int64 SeriesCatalog { get; set; }
     }
 }
